requirements.txt
Flask
flask-cors
pandas
numpy
tensorflow-cpu
scikit-learn
sentence_transformers
mysql-connector-python  
